Controllers:: añadida autorizacion de las policies

+++ app/Http/Controllers/Api/CategoryController.php
@@ -25,17 +25,20 @@ class CategoryController extends Controller
 
     public function show(Category $category)
     {
+        $this->authorize('view', $category);
         return new CategoryResource($category->load('products'));
     }
 
     public function update(CategoryRequest $request, Category $category)
     {
+        $this->authorize('update', $category);
         $category->update($request->except('products'));
         return response()->json(['message' => 'Updated successfully'], 200);
     }
 
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
         $category->delete();
         return response()->json(['message' => 'Deleted successfully'], 200);
     }
